Refactor game into scene-based loading
It is partially done: the main menu (currently automatically) loads the
next scene, but activation of the structural game objects (camera, gui,
gamecontroller etc.) does not want to work properly.

Instantiating those in a disabled state and then enabling them ourselves
from the editor seems to work reasonably well, but it is not a permanent
solution.